+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.2 (unreleased)
 
+- Added `predict` method
 - Fixed `item_ids` option for `user_recs`
+- Fixed `user_recs`
 
 ## 0.2.1 (2020-10-28)